scripts/requirements_cxr_foundation.txt
numpy==1.26.4
Pillow>=10,<12
pypng>=0.20220715.0
huggingface_hub>=0.28,<1
tensorflow==2.17.1
tensorflow-text==2.17.0
